Move Board component to seperate jsx
